// apps/web/src/data/seed.generated.ts
export const SEED_EN: WordEntry[] = [
 {
  "id": "w_mssmlovd_alaslc",
  "word": "approve",
  "tags": [
   "công việc"
  ],
  "dateAdded": "2026-08-14T07:28:52.009Z",
  "learningStatus": "new",
  "correctStreak": 0,
  "synonyms": [
   "agree",
   "sanction",
   "endorse"
  ],
  "antonyms": [
   "disapprove",
   "reject",
   "deny"
  ],
  "wordRoot": "ad (to) + probare (to test, approve)",
  "senses": [
   {
    "pronunciation": "/əˈpruːv/",
    "partOfSpeech": "verb",
    "meaning": {
     "en": "To officially agree to or accept something as satisfactory.",
     "vi": "Đồng ý, chấp thuận."
    },
    "examples": [
     "The committee approved the new budget."
    ]
   }
  ]
 },
// ...
];

export const SEED_ZH: WordEntry[] = [
 {
  "id": "w_mssmlovj_45xer9",
  "word": "你好",
  "tags": [
   "giao tiếp"
  ],
  "dateAdded": "2026-08-14T07:28:52.015Z",
  "learningStatus": "new",
  "correctStreak": 0,
  "synonyms": [],
  "antonyms": [],
  "wordRoot": "",
  "senses": [
   {
    "pronunciation": "nǐ hǎo",
    "partOfSpeech": "interjection",
    "meaning": {
     "vi": "Xin chào."
    },
    "examples": [
     "你好，我叫安娜。"
    ]
   }
  ]
 },
// ...
];